triposr-server/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
rembg>=2.0.57
trimesh>=4.2.0
# TripoSR itself — install from source (see setup.sh)
# huggingface_hub is pulled in by tsr
